requirements.txt
build

#test
pytest
pyserial